+++ src/pages/onboarding/OnBoardingPage.jsx
@@ -1,0 +1,12 @@
+import * as S from './OnBoardingPage.style';
+import OnBoardingCardPage from '@/components/onboarding/OnBoardingCard';
+
+const OnBoardingPage = () => {
+  return (
+    <S.SlickWrapper>
+      <div>안녕</div>
+      <OnBoardingCardPage />
+    </S.SlickWrapper>
+  );
+};
+export default OnBoardingPage;
