+++ src/pages/onboarding/OnBoardingPage.style.js
@@ -1,0 +1,14 @@
+
+import styled, { css } from 'styled-components';
+
+
+const SlickWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  align-items: center;
+  padding-bottom: 40px;
+  text-align: center;
+`;
+
+export {SlickWrapper  };
